+++ CHANGELOG.md
@@ -15,8 +15,11 @@
 - DAPI-739: Add coefficients by criterion to the calculation of the axes rates
 - DAPI-635: Add spellcheck on WYSIWG editors
 - DAPI-798: Allow spelling suggestions after a title formatter check
-- RUL-20: Rule engine - As Julia, I would like to copy values from/to different attribute types
 - DAPI-895: As Julia, I'd like spell-check to be available for Norwegian
+- DAPI-749: Improve Dashboard rates purge
+- RUL-20: Rule engine - As Julia, I would like to copy values from/to different attribute types
+- RUL-49: Rule engine - As Peter, I would like to clear attribute values, associations, categories and groups
+- RUL-77: Rule engine - As Peter, I would like to add labels to my rules
 
 ## New features